Harden Cloudflare site identity validation
